timestamp fixes and sampling rate update
Fixed how timestamps are handled to be consistently set to UTC
process all timestamps using process_datetime to uniform how timestamps are handled.
output datatime and datestring consistently across output for debugging
added timestamp_s and timestamp_ms in uncompressed landmarker file for debugging
Compute median and mean of sampling rate to catch if there are gaps in sampling in the landmarker files.